Updated merge conflicts, bug fixes from
This branch is not fully working, but it builds and is stable. There are still graphics issues with this where NPC palettes are not loaded properly. This commit is meant to be a step in fixing that glitch.

+++ include/field_player_avatar.h
@@ -3,8 +3,8 @@
 
 void PlayerStep(u8 direction, u16 newKeys, u16 heldKeys);
 void ClearPlayerAvatarInfo(void);
-void SetPlayerAvatarExtraStateTransition(u16, u8);
-u8 GetPlayerAvatarGenderByGraphicsId(u16);
+void SetPlayerAvatarExtraStateTransition(u8, u8);
+u8 GetPlayerAvatarGenderByGraphicsId(u8);
 bool8 TestPlayerAvatarFlags(u8);
 u8 GetPlayerAvatarSpriteId(void);
 void PlayerGetDestCoords(s16 *, s16 *);
@@ -50,8 +50,8 @@ bool8 IsPlayerSurfingNorth(void);
 void SetPlayerAvatarWatering(u8 direction);
 u8 GetPlayerAvatarFlags(void);
 void UpdatePlayerAvatarTransitionState(void);
-u16 GetFRLGAvatarGraphicsIdByGender(u8);
-u16 GetRSAvatarGraphicsIdByGender(u8);
+u8 GetFRLGAvatarGraphicsIdByGender(u8);
+u8 GetRSAvatarGraphicsIdByGender(u8);
 void PlayerWheelieInPlace(u8 direction);
 void PlayerWheelieMove(u8 direction);
 void PlayerPopWheelieWhileMoving(u8 direction);
